Improved_Training_of_Wasserstein_GANs/requirements.txt
matplotlib==3.7.5
numpy==1.24.4
pillow==10.2.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.2